Mostly done
The program now reads the input files, correctly as far as I can tell.
Just need to add a flag to each command to determine whether it should
be run, functions for actually running commands, and the command line
parsing.  Also, need to modify things slightly so that each command has
its proper identifier, and add in program configuration.  Still, the
hard part is done, rest is just tedious.