argv in execvp
Added a couple lines to copy the name of the command into the first
argument of the array of strings passed to execvp.  The function seems
to ignore the first argument, but the manual page says by convention the
first argument should be the name of the program, so I have that copied
just case it matters some how.