Interface
Currently working on interface.  The program reads the file and executes
the commands perfectly as far as I can tell.  Just need to have it read
a configuration file, add option to pass file on the command line, and
give it the ability to read all scripts in a specified directory.