Mostly Done
The program reads from the config directory correctly, and is able to
get scripts in the appropriate directory.  Parses command line options,
and seems to work correctly.  Fully functional, just want to clean up
some of the code, have better error messages and more information while
running.  Also, need to improve the dictionary, since it currently uses
linear probing.

Many function prototypes moved to main.h, since they were cluttering up
main.c, and don't really belong in a separate module.  I have also added
comments to most of the main function, but I need to comments the rest
of the functions in the other modules.  Also, need a README and a proper
makefile.  Will see later about using the autotools.  Probably overkill
for something like this, but would make installation easier, as
otherwise users would have to manually move the binary to path, and
remove it if they choose to uninstall it.  Also, need to find out if it
handles manual pages.