Comments and optind
Added a few comments.  Also, if no ids are passed on the command line,
then *all* commands are marked, either to be shown, or in a dry-run.
The to_execute flag is set to zero in this case to prevent the user from
unintentially running all commands.  There should probably be a flag to
force execution of everything, but most of the time, that's probably not
what was intended.