Completion and main
Providing ids with no flags now results in the program printing the
arguments and exiting.  If nothing is passed, all commands are shown.

The completion script has been altered to actually be useful.  It will
show ids by default, otherwise if previous argument is --file, will show
files in current directory, and current argument is --* it will complete
options.